+++ Models/Company.swift
@@ -1,0 +1,16 @@
+//
+//  Model.swift
+//  Typicode
+//
+//  Created by Кирилл Кибешев on 21.01.2020.
+//  Copyright © 2020 Кирилл Кибешев. All rights reserved.
+//
+
+import Foundation
+
+struct Company: Codable {
+    let name: String?
+    let catchPhrase: String?
+    let bs: String?
+}
+
