+++ Models/Geo.swift
@@ -1,0 +1,14 @@
+//
+//  Geo.swift
+//  Typicode
+//
+//  Created by Кирилл Кибешев on 22.01.2020.
+//  Copyright © 2020 Кирилл Кибешев. All rights reserved.
+//
+
+import Foundation
+
+struct Geo: Codable {
+    let lan: String?
+    let lng: String?
+}
