--- Typicode/Model.swift
@@ -1,40 +1,0 @@
-//
-//  Model.swift
-//  Typicode
-//
-//  Created by Кирилл Кибешев on 21.01.2020.
-//  Copyright © 2020 Кирилл Кибешев. All rights reserved.
-//
-
-import Foundation
-
-struct User: Codable {
-    let id: Int?
-    let name: String?
-    let username: String?
-    let email: String?
-    let phone: String?
-    let website: String?
-    let address: Address?
-    let company: Company?
-}
-
-struct Address: Codable {
-    let street: String?
-    let suite: String?
-    let city: String?
-    let zipcode: String?
-    let geo: Geo?
-}
-
-struct Geo: Codable {
-    let lan: String?
-    let lng: String?
-}
-
-struct Company: Codable {
-    let name: String?
-    let catchPhrase: String?
-    let bs: String?
-}
-
