+++ Typicode/ViewController.swift
@@ -11,22 +11,20 @@ import UIKit
 class ViewController: UIViewController {
 
     private let manager = UsersAPIManager()
-    private let url = "https://jsonplaceholder.typicode.com/users"
     private let tableView = UITableView()
     private var users: [User] = []
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        self.view.addSubview(tableView)
-        tableView.translatesAutoresizingMaskIntoConstraints = false
         configureTableView()
         loadData()
         tableView.delegate = self
         tableView.dataSource = self
     }
 
-    func loadData() {
+    private func loadData() {
+        let url = "https://jsonplaceholder.typicode.com/users"
         manager.getUsers(urlString: url, completion: { [weak self] users in
             guard let self = self else {
                 return
@@ -39,7 +37,10 @@ class ViewController: UIViewController {
 
     }
 
-    func configureTableView(){
+    private func configureTableView() {
+
+        self.view.addSubview(tableView)
+        tableView.translatesAutoresizingMaskIntoConstraints = false
 
         tableView.leftAnchor.constraint(equalTo: view.leftAnchor, constant: 0).isActive = true
         tableView.topAnchor.constraint(equalTo: view.topAnchor, constant: 0).isActive = true
@@ -62,8 +63,7 @@ extension ViewController: UITableViewDataSource {
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = UserTableViewCell()
         let user = self.users[indexPath.row]
-        cell.nameLabel.text = user.name
-        cell.phoneLabel.text = user.phone
+        cell.configure(user: user)
         return cell
     }
 }
